+++ CHANGELOG.md
@@ -9,24 +9,34 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## 📦 [Unreleased] 
 
+### 💎 Added
+
+- Methods:
+  - `isEmpty` | `isNotEmpty`
+
 ### ⛏️ Work in progress
 
 - Methods:
+  - `isBlank` | `isNotBlank`
   - `isAlphaDigit` | `isNotAlphaDigit`
 
 ### 💡 Ideas for future release
 
 - `formatNumber(string): string` (replace '.' by ',')
-- `wordsToArray(string): string[]`: returns the words of a sentence in an array.
-    Example : `wordsToArray('Hello the world') => ['Hello', 'the', 'world']`.
+- `containsUpperCase`: contains at least one uppercase letter.
+- `containsLowerCase`: contains at least one lowercase letter
+- `isStrongPassword`: string that contains at least 1 uppercase,
+  1 lowercase, 1 number, 1 special character and has a minimum length of 12 characters.
 - `toNumber(string): number`: parse string to number.
 - `toBoolean(string): boolean`: parse string to boolean (case insensitive).
+- `wordsToArray(string): string[]`: returns the words of a sentence in an array.
+    Example : `wordsToArray('Hello the world') => ['Hello', 'the', 'world']`.
+- `removeWords`: remove list of words from a given string.
 - Case style:
   - `camelCase`: convert string to camelCase.
   - `upperCamelCase`: convert string to UpperCamelCase.
   - `snakeCase`: convert string to snake_case.
   - `kebabCase`: convert string to kebab-case.
-- `removeWords`: remove list of words from a given string.
 
 ---