+++ CHANGELOG.md
@@ -15,15 +15,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `isEmpty` | `isNotEmpty`
   - `isBlank` | `isNotBlank`
   - `isAlphaNumber` | `isNotAlphaNumber`
+  - `containsUpperCase`
+  - `containsLowerCase`
 
 ### ⛏️ Work in progress
 
 - Methods:
-  - `containsUpperCase`: contains at least one uppercase letter.
-  - `containsLowerCase`: contains at least one lowercase letter
   - `isStrongPassword`: string that contains at least 1 uppercase,
-  - `formatNumber(string): string` (replace '.' by ',')
     1 lowercase, 1 number, 1 special character and has a minimum length of 12 characters.
+  - `formatNumber(string): string` (replace '.' by ',')
 
 ### 💡 Ideas for future release